moved phase data to another file
