+++ LuckParser/LuckParser.csproj
@@ -89,6 +89,7 @@
     <Compile Include="Models\ParseModels\AbstractPlayer.cs" />
     <Compile Include="Models\ParseModels\Boons\BoonsGraphModel.cs" />
     <Compile Include="Models\ParseModels\Boons\BoonDistribution.cs" />
+    <Compile Include="Models\ParseModels\PhaseData.cs" />
     <Compile Include="Models\ParseModels\Simulator\BoonSimulationItem.cs" />
     <Compile Include="Models\ParseModels\Boss.cs" />
     <Compile Include="Models\ParseModels\GW2APISpec.cs" />
